Expand order class, include tests, import authentication module

+++ checkout-microservice/src/main/java/nl/tudelft/sem/checkout/domain/OrderNotFoundException.java
@@ -3,7 +3,7 @@ package nl.tudelft.sem.checkout.domain;
 public class OrderNotFoundException extends Exception{
     static final long serialVersionUID = -3387516993124229948L;
 
-    public OrderNotFoundException(int orderId) {
+    public OrderNotFoundException(long orderId) {
         super(String.valueOf(orderId));
     }
 }
