+++ checkout-microservice/src/main/java/nl/tudelft/sem/checkout/domain/OrderService.java
@@ -19,7 +19,7 @@ public class OrderService {
         return orderRepository.findAll();
     }
 
-    public Order getOrderById(int orderId) throws Exception{
+    public Order getOrderById(long orderId) throws Exception{
         if(orderRepository.existsByOrderId(orderId))
             return orderRepository.findByOrderId(orderId).get();
         throw new OrderNotFoundException(orderId);
@@ -45,7 +45,7 @@ public class OrderService {
      * Removes an order from the DB
      * @param orderId - the id of the Order object to be removed
      */
-    public void removeOrderById(int orderId) {
+    public void removeOrderById(long orderId) {
         orderRepository.deleteOrderByOrderId(orderId);
     }
